# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ÖzelGüvenlik.Online API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
  /users/me:
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        nameColor:
          type: ["string", "null"]
        nameAnimated:
          type: boolean
        isBanned:
          type: boolean
        banReason:
          type: ["string", "null"]
        banExpiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
    UserProfile:
      type: object
      required: [id, username, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        nameColor:
          type: ["string", "null"]
        nameAnimated:
          type: boolean
        listingCount:
          type: integer
        createdAt:
          type: string
...
    ProfileUpdate:
      type: object
      properties:
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
...